#6: Rename classes-folder to common
